Modification minor
Uncommmented the part that is dividing the validation and training sets. This time smaller part of the data is not planned to be used by the validation set (since the cross validation option of the fitcsvm is selected), but this smaller portion is used for the test set.

Added some figure outputs from code